+++ l10n_br_account_product/models/l10n_br_account_product_icms_relief.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2013  Renato Lima - Akretion
+# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
+
+from openerp import models, fields
+
+
+class L10nBrIcmsRelief(models.Model):
+
+    _name = 'l10n_br_account_product.icms_relief'
+    _description = 'Icms Relief'
+
+    code = fields.Char(u'Código', size=2, required=True)
+    name = fields.Char('Nome', size=256, required=True)
+    active = fields.Boolean('Ativo', default=True)
